Slight files re-arrangements and renames
